requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
psycopg-binary==3.1.17
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==8.7.0
typing_extensions==4.9.0
urllib3==2.2.1
